Modularize Lottie renderer evaluators and layer composables into domain packages
### Impact Summary
Decomposes monolithic Lottie animation evaluators and layer composables into domain-specific renderer subpackages (`renderer.properties`, `renderer.shapes`, and `renderer.layers`). This structural refactoring introduces no behavior or runtime changes and maintains complete functional parity across all animation rendering, interpolation, and screenshot verification tests.

### Motivation
Previously, animation property evaluators (`animateScalar`, `animatePosition`, `animateVector`, `animateColor`, `animateBezier`), shape converters (`path`, `rectangle`, `ellipse`), and layer composables (`Layer`, `ShapeLayer`, `SolidColorLayer`) were embedded inside large monolithic files (`Animation.kt`, `Shape.kt`, `Layer.kt`). Extracting them into fine-grained domain packages isolates individual rendering concerns, facilitates focused unit testing, and provides the foundation for upcoming feature additions such as hold keyframes, LUT easing, shape modifiers, and matte clipping.
