+++ remotecompose/lottie/src/main/java/com/google/android/horologist/remotecompose/lottie/renderer/properties/Color.kt
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2026 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.android.horologist.remotecompose.lottie.renderer.properties
+
+import android.annotation.SuppressLint
+import androidx.compose.remote.creation.compose.state.RemoteColor
+import com.google.android.horologist.remotecompose.lottie.LottieSettings
+import com.google.android.horologist.remotecompose.lottie.format.properties.StaticColorProperty
+
+/**
+ * Resolves or animates a color property.
+ *
+ * Checks for slot override in [LottieSettings.slotMap] before falling back to property value.
+ */
+@SuppressLint("RestrictedApi")
+internal fun animateColor(
+  property: StaticColorProperty,
+  animationSettings: LottieSettings,
+): RemoteColor {
+  val slotColor = property.slotId?.let { animationSettings.slotMap.getColor(it) }
+  return slotColor ?: property.value
+}
