+++ remotecompose/lottie/src/test/java/com/google/android/horologist/remotecompose/lottie/AnimationTest.kt
@@ -21,7 +21,7 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.android.horologist.remotecompose.lottie.format.properties.AnimatedVectorProperty
 import com.google.android.horologist.remotecompose.lottie.format.properties.StaticVectorProperty
 import com.google.android.horologist.remotecompose.lottie.format.properties.VectorPropertyKeyframe
-import com.google.android.horologist.remotecompose.lottie.renderer.animateVector
+import com.google.android.horologist.remotecompose.lottie.renderer.properties.animateVector
 import com.google.common.truth.Truth.assertThat
 import org.junit.Test
 import org.junit.runner.RunWith
