+++ remotecompose/lottie/src/test/screenshots/MediaLottieDiffScreenshotTest_m3Next.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:d4df5565d3127c78c1c9fa54eb654710e9fb561e590e02c585587481357fc270
-size 12330
+oid sha256:16b03f5373d15f8af05699963faa34a4898a384d3d662d16ae18042ffd1fc87f
+size 12587
