+++ remotecompose/lottie/src/test/screenshots/MediaLottieDiffScreenshotTest_muteToUnmute.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:e129dae8e79144a859ede80525499b4b1ea43fc4d3bb804c0e9da55b45411c6b
-size 14577
+oid sha256:e8b140622400a14ce8ed52d99817732643c1991aeb51ce22c1db2bbd2dcff75d
+size 14843
