+++ remotecompose/lottie/src/test/screenshots/MediaLottieDiffScreenshotTest_volumeUp.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:d5469e889e9f13bab1efe7a32f4477641514fd436c06a75c384211b08f879e0b
-size 14474
+oid sha256:67edb350f02d1ebcf70e85d15ee2e6cc828bb9ceadfe8764b1fca7fb5ccd62a3
+size 14741
